Home Page Desktop Layout
Here I've finished the main layout for the home page.

I'm now working on the blog page for the website. As before, I've laid-out the page with grid changing the background colour of each section. This is to show where each section on the page is. I'll continue to work within each section's container.