Home Page Media Queries
Here I've added some media queries to change the layout of the page on smaller screens.

As the screen gets smaller, the useful links section is removed and fewer blogs are displayed in one line across the width of the screen. On tablet 2 blog cards take up the full width, and on mobile one blog card takes up the full screen width.

I'll continue to work on media queries in the next commit.