About The Author section change
Here I've changed the layout of the About The Author section.

I've also changed the nav bar to be sticky, and it now includes a home button and a picture of the author.

The desktop layout for this page is now finsihed. In the next commit I'll be working on making it mobile responsive.